// src/app/generated-apis/user/model/gender.ts
/**
 * user-generated-contract
 * user-generated-contract
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Gender = 'MALE' | 'FEMALE';

export const Gender = {
    Male: 'MALE' as Gender,
    Female: 'FEMALE' as Gender
};