// src/app/generated-apis/user/model/workingDays.ts
/**
 * user-generated-contract
 * user-generated-contract
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type WorkingDays = 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';

export const WorkingDays = {
    Monday: 'MONDAY' as WorkingDays,
    Tuesday: 'TUESDAY' as WorkingDays,
    Wednesday: 'WEDNESDAY' as WorkingDays,
    Thursday: 'THURSDAY' as WorkingDays,
    Friday: 'FRIDAY' as WorkingDays,
    Saturday: 'SATURDAY' as WorkingDays,
    Sunday: 'SUNDAY' as WorkingDays
};